+++ docs-app/lib/ui-merged.js
@@ -100,6 +100,7 @@ export const {
   Button,
   ButtonGroup,
   Card,
+  CardChange,
   CardHeader,
   CardTitle,
   CardDescription,